+++ src/min_client.py
@@ -16,8 +16,7 @@ import sounddevice as sd
 from websockets.sync.client import connect
 import numpy as np
 
-from modules.packet_conv.audio import enc_from_ndarray, dec_to_ndarray
-import AUDIO_PARAM
+from modules.packet_conv import AUDIO_PARAM, audio_ndarray
 
 
 if __name__ == "__main__":
@@ -29,10 +28,10 @@ if __name__ == "__main__":
   ws = connect(SERVER_URI)
 
   def callback(indata: np.ndarray, outdata: np.ndarray, frames, time, status):
-    packet_send = enc_from_ndarray(indata, bytes())
+    packet_send = audio_ndarray.encode(indata, bytes())
     ws.send(packet_send)
     packet_recv = ws.recv()
-    outdata[:], _ = dec_to_ndarray(packet_recv)
+    outdata[:], _ = audio_ndarray.decode(packet_recv)
 
 
   try:
